fix popover background
