+++ Sources/Playbook/Resources/Extensions/View+CancelFirstResponder.swift
@@ -9,25 +9,28 @@
 
 import SwiftUI
 
-struct CancelFirstResponder: ViewModifier {
-  func body(content: Content) -> some View {
-    content
-      .onTapGesture {
-        #if os(iOS)
-        UIApplication.shared.sendAction(#selector(UIResponder.resignFirstResponder), to: nil, from: nil, for: nil)
+//struct CancelFirstResponder: ViewModifier {
+//  let action: (() -> Void)
+//
+//  func body(content: Content) -> some View {
+//    content
+//      .onTapGesture {
+//        action()
+//        #if os(iOS)
+//        UIApplication.shared.sendAction(#selector(UIResponder.resignFirstResponder), to: nil, from: nil, for: nil)
 //        #elseif os(macOS)
 //        NSEvent.addLocalMonitorForEvents(matching: .leftMouseUp) { event in
 //          event.window?.makeFirstResponder(nil)
 //          event.window?.makeFirstResponder(event.window)
 //          return event
 //        }
-        #endif
-        }
-  }
-}
-
-public extension View {
-  func cancelFirstResponder() -> some View {
-    self.modifier(CancelFirstResponder())
-  }
-}
+//        #endif
+//        }
+//  }
+//}
+//
+//public extension View {
+//  func cancelFirstResponder(action: @escaping (() -> Void)) -> some View {
+//    self.modifier(CancelFirstResponder(action: action))
+//  }
+//}
